services/ai-agent-service/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-generativeai==0.3.2
requests==2.31.0
gunicorn==21.2.0
nltk==3.8.1
langchain==0.0.350
langchain-google-genai==0.0.6
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.1
eventlet==0.33.3

